bugs fixed

+++ backend/routers/category.router.js
@@ -17,6 +17,7 @@ router.post("/add", async (req, res) => {
       const category = new Category({
         _id: uuidv4(),
         name: name,
+        isActive: false,
       });
 
       await category.save();
@@ -28,8 +29,17 @@ router.post("/add", async (req, res) => {
 router.post("/removeById", async (req, res) => {
   response(res, async () => {
     const { _id } = req.body;
-    await Category.findByIdAndDelete(_id);
-    res.json({ message: "Kategori kaydı başarı ile silindi!" });
+    const category = await Category.findById(_id);
+    if (category.isActive === false) {
+      await Category.findByIdAndDelete(_id);
+      res.json({
+        message: `Kategori kaydı başarı ile silindi!`,
+      });
+    } else {
+      res.json({
+        message: `Kullanılan kategori silinemez!`,
+      });
+    }
   });
 });
 
@@ -57,6 +67,21 @@ router.get("/", async (req, res) => {
   response(res, async () => {
     const categories = await Category.find().sort({ name: 1 });
     res.json(categories);
+  });
+});
+
+router.post("/changeStatus", async (req, res) => {
+  response(res, async () => {
+    const { _id, isActive } = req.body;
+
+    let category = await Category.findById(_id);
+
+    category = {
+      isActive: isActive
+    }
+
+    await Category.findByIdAndUpdate(_id, category);
+    res.json({ message: "Kategori durumu değişti" });
   });
 });
